Fix pytorch version and git repo information logged
Note that the git repo is not always the pytorch repo, so
add a repo name field, and rename all the repo fields to
reflect they are generic.

Pytorch version is logged inside the uploader, which is
potentially dangerous if the uploader is not used from
the same environment as the benchmark, althrough it currently is.